scrapy4demo/requirements.txt
astroid==2.3.3
attrs==19.3.0
automat==20.2.0
bcrypt==3.1.7
cffi==1.14.0
constantly==15.1.0
cryptography==2.9
cssselect==1.1.0
demjson==2.2.4
deprecated==1.2.9
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.0 ; python_version != '3.4'
mccabe==0.6.1
paramiko==2.7.1
parsel==1.5.2
protego==0.1.16
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pydispatcher==2.0.5
pyhamcrest==2.0.2
pylint==2.4.4
pymongo==3.10.1
pynacl==1.3.0
pyopenssl==19.1.0
queuelib==1.5.0
scrapy-proxies==0.4
scrapy-random-useragent==0.2
scrapy-splash==0.7.2
scrapy==2.0.1
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
sshtunnel==0.1.5
twisted==20.3.0
typed-ast==1.4.1 ; implementation_name == 'cpython' and python_version < '3.8'
urllib3==1.25.9
w3lib==1.21.0
wrapt==1.11.2
yapf==0.29.0
zope.interface==5.1.0